+++ src/components/currency/Currency.tsx
@@ -1,6 +1,6 @@
 import "./Currency.css";
 
-var formatter = new Intl.NumberFormat("en-US", {
+const formatter = new Intl.NumberFormat("en-US", {
   style: "currency",
   currency: "USD",
   maximumFractionDigits: 2,
